docopt: fix optargs
Optargs should not be prepended with the same rule name.
For example:

Usage: program [<name>]

name:
    | name WORD

above is incorrect, since the following can be also accepted:

"programe name1 name2 name3 ..."

What we want is strict:

"program name1"

So the rules should be generated as following:

name:
    | WORD

And for me still is not clear why bison complains with warning.
But for now we do not care.